define script life cycle in redis
before this commit, we loaded scripts to redis in every cold start. This is not so desirable because the scripts in the redis cache stay there unless someone calls SCRIPT FLUSH. Hence, a life cycle is needed.

How do we define a life cycle for a serverless environment? Random numbers. In each limit() invocation, we create a random number and flush the scripts if it is smaller than some frequency threshold. If a user makes a lot of requests, their scripts will get flushed eventually.

This implementation however, has a hazard: serverless instance A can flush while instance B is still operating. We need a 'safeEvalsha' and we add exactly that in hash.ts.